Avoid shadowing the length function
I'm not entirely sure what's happening here but Emacs 26 fails to expand this method definition if the `length` variable shadows the `length` function. That is, if you don't do this, Emacs substitutes all occurrences of `length` with `(slot-value obj 'length)` and then complains about this being an invalid function (when it tries to compile `((slot-value object 'length) tmp)`.

I'm pretty sure this is a bug in emacs.